setup.py: Fix linking issues on Unix systems. Fixes #1787
When building with local MuPDF, we should not tell setuptools to use link args
such as '-l harfbuzz', because all required symbols will already be in libmupdf
or libmupdf-third.

In fact specifying args like '-l harfbuzz' can actually cause an error if the
library is not installed on the system, even though it is not needed to resolve
any symbols.

Also added some minor diagnostics.

So this commit removes '-l' and associated '-L' linker args when using a local
MuPDF build.