fitz/helper-other.i: Archive support: added JM_last_tree() and JM_archive_from_py().
Function "JM_last_tree":

The "arch" parameter must be a multi-archive. This functions returns either
NULL or the last sub-archive of "arch" if it is of type "tree" and has been
added with the same "mount" value as in the arguments.  Thus this functions
avoids bloating the multi-archive with single-element sub-archives added in a
row.

Function "JM_archive_from_py"
Is no longer used and can be removed!