Fixes to optionally work with MuPDF master's new filter api.
If mupdf is a git checkout, we run a git command to find
whether it is master branch. This can be overridden by setting
PYMUPDF_SETUP_MUPDF_BUILD_BRANCH=master when running setup.py.

If MuPDF is master branch, we predefine MUPDF_BRANCH_master when building,
which enables code which uses the new MuPDF API.

setup.py: also: removed trailing '/' from `mupdf_local` so that code has to use
explicit '/' characters when using it.