fitz/: Fix #2238 - use 'overlap' rather than 'contains' when extracting text.
We now include chars that overlap with the clipbox, instead of only those that
are entirely contained within the clipbox.

Note that new fn JM_rects_overlap() still returns true if one of the rects is
empty. This allows things to work with ligatures, where component glyphs can
have zero width.