Fix #2246
`TextWriter.write_text` used to ignore cases where MediaBox (x0, y0) != (0, 0). Similarly, in cases of page rotation being equal to 90 or 270, the positioning of text on the page failedto take this into account.
Note that these adjustments needed to be done in method `write_text` because a `TextWriter` can be written to multiple pages, which might have different CropBox positions or page rotations.