Revert all changes to workflows made on branch 1.21 since it branched off master.
This is the first stage of moving all PyMuPDF development to a new branch
`main`, and mothballing branch 1.21.

This commit reverts the small number of unused workflow-related commits that
were made on branch 1.21 since it branched off master, so that we can later
trivally cherry-pick the more useful workflow-related commits from master and
rename to new branch `main`.