fitz/utils.py src/utils.py: Re-apply fix for #2345 that was unintentionally reverted.
Previously we displayed error messages (Python `print()`) if a link could not
be replicated on the target page within method `doc.insert_pdf()`.

As the method has no way to provide more information on the problem's cause,
and hence cannot provide information on how to fix the problems, the messages
make no sense and will be omitted.