Enhancement: Correctly setting RadioButton widgets
We did not previously sufficiently handle correctly setting radio buttons to "ON" in an RB group. This fix does the following:
* Setting on can be achieved by setting the field value to `True` or the name of the "ON" state of the object
* If `True` is used, the correct PDF name value is extracted via MuPDF function, else the provided value is accepted
* The other buttons in the RB group are identified and set to `/Off` using PyMuPDF's low-level interface (Python)
* MuPDF is invoked to set the button to ON.

There also is a new method `Widget.on_state()` which returns the name of the buttons ON state as a string. This value can alternatively be used to correctly setting on the button: `widget.field_value = widget.on_state()`.
The method can also be used to check the state: `widget.field_value == widget.on_state()` being true means that the button is checked. The alternative is checking `widget.field_value != "Off"`.

fitz/helper-fields.i