Support variable border width
For writing text with a glyph border, render mode Tr > 0 must be used.
Parameter "border_width" controls the thickness of the glyph border.

This fix adds support for "border_width" to TextWriter.
This parameter is interpreted as a float factor to be multiplied with the fontsize. Useful values are in the range of 0.05 (5% of the fontsize, default) or less.
This interpretation is now also consistent with "insert_text()" / "insert_textbox()" of the Page and Shape class.