Fix #2637
Correctly compute the filled area (or overflow) of method Shape.insert_textbox.

Also support variable setting of character border widths for insert_text() / insert_textbox(). This is a factor to be multiplied with the font size. Default is 0.05 (read: 5% of the fontsize). This value is relevant for text rendering modes 1 and 2 only.

Prepare "words! extraction for variable delimiters. This is only functional until after changes in __init__.py.