rebase stage 2: rename fitz/ to fitz_old/.
Unlike with rebased, renaming the generated classic Python module is quite
tricky, because the SWIG-generated code is called `fitz` and things do not work
if we simply rename the pure-python code to be `fitz_old`.

So this is the first of a series of commits that renames all classic code to
generate a module called `fitz_old`.