docs/page.rst: various cleanups to history items.
Moved all history comments to end of relevant line or section.

All end-of-methods historical info is hidden by default.

Inline historical info for individual args is shown all the time as before, but
moved to the end of the relevant sentence or paragraph, so it's not the first
thing one reads

Uses docs/header.rst's new |history|...|history_end| to allow toggle of history
info.