Re-Implement select method
This re-implements Document method "select()" based on new MuPDF function "pdf_rearrange_pages()".
This is a more complete (and faster) implementation of what needs to be done here in that not only pages will be rearranged, but also consequential changes will be made to the table of contents, links to removed pages and affected entries in the Optional Content definitions.

Update __init__.py