setup.py: Only overwrite MuPDF's include/mupdf/fitz/config.h on Windows.
This avoids modifying a local MuPDF tree when developing. There is no need on
non-Windows because we already predefine TOFU_CJK_EXT in XCFLAGS. On Windows
there is no equivalent to XCFLAGS so for simplicity we continue to overwrite
MuPDF's config.h.

For completeness we now also predefine TOFU_CJK_EXT in XCXXFLAGS for the C++
bindings compiles, although in practise this will not make any difference.

Also modify env_add() to default to adding to os.environ[name] if name is not
in env, so we always append to existing environment variable if available.