More permissive URI links
We previously only accepted external links as LINK_URI if the "uri"-string started with certain prefixes.
With this change /URI links are turned into LINK_URI links if they contain a colon(":") and do not start with prefix "file:".
In all other cases, a LINK_LAUNCH will be generated.

This change also improves the recognition of TOC LAUNCH and GOTOR destinations, making changes in the test scripts necessary.

We only test this on rebased and we no longer differentiate by MuPDF or PyMuPDF versions.