Find end-of-text marker correctly
When modifying annotation appearance streams, finding the PDF end-of-text marker b"ET" must be the last one in the stream. I.e. use rfind(b"ET") instead of find(b"ET").
We erroneously searched for the next one - which can destroy the stream if it happens to contain multiple b"ET" strings.