Add new page method "remove_rotation"
The new method sets a page's rotation to 0 while keeping the page's appearance and content unchanged.
This can significantly simplify inclusion of the page in target pages via Page method "show_pdf_page()".

Other changes are additions / corrections to "changes.txt" and improved wordings for in document.rst and page.rst.

Also contained is a fix for issue 3402 (inability to insert/update fields with inter-field calculations.