.github/workflows/build_wheels.yml: use macos-13 and macos-14.
This fixes generation of MacOS x86/arm64 wheels.

Actually macos-13 appears to build both x86_64 and arm64 wheels, presumably the
latter using cross-compilation; these arm64 wheels appear to be the ones that
appear in the final artifact (maybe because they are created later).

But macos-13 does not run pytest on the arm64 wheels, so we continue to also build on macos-14.