Correctly locate tessdata from installation folder
Funtionality cannot be tested in GitHub actions because no Tesseract is installed.

If tesseract-ocr is not version 4.00 in a Unix-like platform, we incorrectly used pathlib.Path for locating tessdata.
This fix correctly determines the tessdata folder name for any version of tesseract-ocr.
Note however, that we cannot guarantee that MuPDF's OCR code (which is on version 4.00) can cope with tessdata content contained in a different version than 4.00.

Update __init__.py