Ensure that words never contain RTL/LTR char mixtures
Previously, we only used white space to perform  word breaks. This could lead to mixtures of RTL and LTR characters in the same word string.
Which in turn made it impossible to produce satisfying extractions of RTL / LTR text mixtures.
This change ensures that every word string contains either no or only RTL characters.

Additional standard word delimiters:

0x202A: LEFT-TO-RIGHT EMBEDDING
0x202B: RIGHT-TO-LEFT EMBEDDING
0x202C: POP DIRECTIONAL FORMATTING
0x202D: LEFT-TO-RIGHT OVERRIDE
0x202E: RIGHT-TO-LEFT OVERRIDE

Word breaks will be generated at the occurrence of any of these characters.
In addition, breaks are also made if characters in a row are not both, either RTL or LTR.