Remove proprietary glyphname <> unicode conversion
PyMuPDF unnecessarily provides proprietary code for Unicode to glyph name conversion and back.
This fix uses the built-in Python module "unicodedata" for providing the same functionality.
Note that we import this module within the respective two functions. They are rarely used and thus this import shouldn't charge PyMuPDF's import with additional processing.