Allow arbitrary ON values for CheckBoxes
We previously insisted that the ON-values for CheckBoxes are defined as "Yes" in the PDF (the published recommended value).

This fix introduces lifts this restriction and supports arbitrary values as ON for  check boxes.

So a check box turns to OFF if setting its field value to `False` or "Off".
It is set to ON if the field value is set to any of 'True`, "Yes" or the respective value coded in the PDF.

When checking a check box's value, the return values invariably are either "Off" or the ON-value in the PDF. Which value this is, can be checked via method `on_state()`.