tests/: added stext bbox test, test_4179.
This shows problem with square root symbol font definition, which requires
pymupdf.TEXT_ACCURATE_BBOXES and disabling of our quad_corrections as a
workaround.

This new test is disabled with non-optimised pymupdf - looks like behaviour
differs because of Pythons's use of unicode instead of raw utf8, and it's
non-trivial to fix.