Remove dependency on MuPDF version
Add all MuPDF STEXT flags up to v1.26.0 to PyMuPDF.
Use hard coded values if unknown in an earlier MuPDF version that we still want / need to support.
The intention is to switch to MuPDF's symbolic names as soon as we drop support of the corresponding version.
Flag bits representing current MuPDF features can always be used because the are ignored by older MuPDF versions.

Also removed some duplicate definitions.