scripts/test.py: extend and fix --cibw-test-project with pyodide.
In build() with pyodide, use python-3.13 to make setjmp/longjmp work.

In cibw_do_test_project(), add C++ library libqwerty.so, to make test more
similar to PyMuPDF. With Pyodid, it seems that we need to put libqwerty.so in
top level of wheel, not in <package>.libs/ as we did with Python-3.12.

Changed some print(...) statements to log(...).

On Pyodide don't add verbose `-s` flag to pytest flags.