Revert version change for MuPDF 1.26.11 since the causing commit was reverted.
Originally MuPDF 1.26.11 was planned to include both the table hunting code
as well as dehyphenation and the new text search interface. This combination
of changes caused new behaviour in MuPDF which was adapted for in PyMuPDF
commit 23de71837cc944b4421487628e29c2e1e06a7547.

Due to newly found bugs in the dehyphenation and text search code and
apprehensiveness about adding these many changes to a stable version it was
decided to back these out. That means that the change in PyMuPDF commit
23de71837cc944b4421487628e29c2e1e06a7547 will now fail for MuPDF 1.26.11.

The best is to simply revert the commit, which is exactly what happens here.